requirements.txt
torch
transformers
peft
bitsandbytes
accelerate
scipy
pyyaml
datasets